created route for role deletion, along with controller and service

+++ src/services/role/DeleteRoleService.ts
@@ -2,11 +2,37 @@ import prisma from "../../lib/prisma";
 
 class DeleteRoleService {
   async execute(id: string) {
-    const role = await prisma.role.delete({
+    const roleExists = await prisma.role.findUnique({
       where: {
         id,
       },
     });
+
+    if (!id) {
+      throw new Error("The ID informed is invalid.");
+    }
+
+    if (!roleExists) {
+      throw new Error("Role not found.");
+    }
+
+    const deletedRole = await prisma.role.delete({
+      where: {
+        id,
+      },
+    });
+
+    const roleAfterDeletion = await prisma.role.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    if (roleAfterDeletion) {
+      throw new Error("Failed to delete role.");
+    }
+
+    return deletedRole;
   }
 }
